[0.4.7][PCNET] Fix MSVC2010SP1 x86 dbg compiler warning C4146 CORE-18104
can be observed with MSVC 2010SP1 (16.0.40219.1)
and building in debug configuration for x86 target:
C:\047rls\reactos\drivers\network\dd\pcnet\pcnet.c(1226) : warning C4146: unary minus operator applied to unsigned type, result still unsigned

fix is a partial pick of 0.4.8-dev-589-g a2f7de7ee85024dd4f15ea62f5a6f6340e2890f8

+++ drivers/network/dd/pcnet/pcnet.c
@@ -1223,7 +1223,7 @@ MiniportSend(
   Adapter->CurrentTransmitEndIndex %= Adapter->BufferCount;
 
   Desc->FLAGS = TD1_OWN | TD1_STP | TD1_ENP;
-  Desc->BCNT = 0xf000 | -TotalPacketLength;
+  Desc->BCNT = 0xf000 | -(INT)TotalPacketLength;
 
   NdisMSynchronizeWithInterrupt(&Adapter->InterruptObject, MiSyncStartTransmit, Adapter);
